BACKEND/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26
joblib==1.3.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
# New dependencies for improvements
slowapi==0.1.9  
redis==4.5.4  
celery[redis]==5.3.4  
structlog==23.2.0  
prometheus-client==0.19.0  
aioredis==2.0.1  
asyncio-mqtt==0.16.1  
